requirements.txt
llama-index-core==0.14.10
llama-index-llms-ollama==0.9.0
llama-index-embeddings-ollama==0.8.4
Wikipedia-API==0.8.1
tqdm==4.67.1
python-dotenv==1.0.0
